Build: Password reset tokens saved in the database
