--- auth-server/src/main/java/rodrigues/ferreira/ricardo/auth/server/AuthServer.java
@@ -1,13 +1,0 @@
-package rodrigues.ferreira.ricardo.auth.server;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class AuthServer {
-
-	public static void main(String[] args) {
-		SpringApplication.run(AuthServer.class, args);
-	}
-
-}
