--- auth-server/src/main/java/rodrigues/ferreira/ricardo/auth/server/domain/UserEntity.java
@@ -1,89 +1,0 @@
-package rodrigues.ferreira.ricardo.auth.server.domain;
-
-import org.springframework.data.annotation.CreatedDate;
-
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.EnumType;
-import javax.persistence.Enumerated;
-import javax.persistence.GeneratedValue;
-import javax.persistence.Id;
-import java.time.OffsetDateTime;
-
-@Entity
-public class UserEntity {
-
-	@Id
-	@GeneratedValue
-	private Long id;
-	private String name;
-	@Column(unique = true)
-	private String email;
-	private String password;
-	@Enumerated(EnumType.STRING)
-	private Type type;
-	@CreatedDate
-	private OffsetDateTime createdAt;
-
-	public UserEntity() {
-	}
-
-	public UserEntity(String name, String email, String password, Type type) {
-		this.name = name;
-		this.email = email;
-		this.password = password;
-		this.type = type;
-	}
-
-	public Long getId() {
-		return id;
-	}
-
-	public void setId(Long id) {
-		this.id = id;
-	}
-
-	public String getName() {
-		return name;
-	}
-
-	public void setName(String name) {
-		this.name = name;
-	}
-
-	public String getEmail() {
-		return email;
-	}
-
-	public void setEmail(String email) {
-		this.email = email;
-	}
-
-	public String getPassword() {
-		return password;
-	}
-
-	public void setPassword(String password) {
-		this.password = password;
-	}
-
-	public OffsetDateTime getCreatedAt() {
-		return createdAt;
-	}
-
-	public void setCreatedAt(OffsetDateTime createdAt) {
-		this.createdAt = createdAt;
-	}
-
-	public Type getType() {
-		return type;
-	}
-
-	public void setType(Type type) {
-		this.type = type;
-	}
-
-	public enum Type {
-		ADMIN, CLIENT;
-	}
-}
