--- auth-server/src/main/java/rodrigues/ferreira/ricardo/auth/server/domain/UserRepository.java
@@ -1,10 +1,0 @@
-package rodrigues.ferreira.ricardo.auth.server.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface UserRepository extends JpaRepository<UserEntity, Long> {
-
-	Optional<UserEntity> findByEmail(String username);
-}
