--- auth-server/src/main/java/rodrigues/ferreira/ricardo/auth/server/security/JPAUserDetailsService.java
@@ -1,36 +1,0 @@
-package rodrigues.ferreira.ricardo.auth.server.security;
-
-import rodrigues.ferreira.ricardo.auth.server.domain.UserRepository;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.User;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-
-import java.util.List;
-
-@Service
-public class JPAUserDetailsService implements UserDetailsService {
-
-    private final UserRepository userRepository;
-
-    public JPAUserDetailsService(UserRepository userRepository) {
-        this.userRepository = userRepository;
-    }
-
-    @Override
-    public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
-        final var user = userRepository.findByEmail(email)
-                .orElseThrow(()-> new UsernameNotFoundException(email));
-
-        final var simpleGrantedAuthority = new SimpleGrantedAuthority("ROLE_" + user.getType().name());
-
-        return new User(
-                user.getEmail(),
-                user.getPassword(),
-                List.of(simpleGrantedAuthority)
-        );
-    }
-
-}
