--- auth-server/src/main/java/rodrigues/ferreira/ricardo/auth/server/security/PasswordEncoderConfig.java
@@ -1,16 +1,0 @@
-package rodrigues.ferreira.ricardo.auth.server.security;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.crypto.password.PasswordEncoder;
-
-@Configuration
-public class PasswordEncoderConfig {
-
-	@Bean
-	public PasswordEncoder passwordEncoder() {
-		return new BCryptPasswordEncoder();
-	}
-	
-}
