--- auth-server/src/main/resources/application.yml
@@ -1,24 +1,0 @@
-spring:
-  datasource:
-    driver-class-name: com.mysql.cj.jdbc.Driver
-    url: jdbc:mysql://localhost:3306/user
-    initialize: true
-    username: app_employees
-    password: app_employees
-  jpa:
-    show-sql: 'true'
-    hibernate:
-      ddl-auto: none
-  session:
-    store-type: jdbc
-server:
-  port: 8088
-
-aw:
-  auth:
-    provider-uri: "http://localhost:8082"
-    jks:
-      keypass: 123456
-      storepass: 123456
-      alias: awblog
-      path: keystore/awblog.jks
