--- auth-server/src/test/java/com/algaworks/example/auth/server/AuthServerTests.java
@@ -1,13 +1,0 @@
-package com.algaworks.example.auth.server;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class AuthServerTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
