--- user/src/main/java/rodrigues/ferreira/ricardo/auth/user/api/MyUserRegisterRequest.java
@@ -1,47 +1,0 @@
-package rodrigues.ferreira.ricardo.auth.user.api;
-
-import rodrigues.ferreira.ricardo.auth.user.domain.UserEntity;
-
-import javax.validation.constraints.NotBlank;
-
-public class MyUserRegisterRequest {
-    @NotBlank
-    private String name;
-    @NotBlank
-    private String email;
-    @NotBlank
-    private String password;
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-    public UserEntity toEntity() {
-        return new UserEntity(
-                this.name,
-                this.email,
-                this.password,
-                UserEntity.Type.CLIENT
-        );
-    }
-}
