--- user/src/main/java/rodrigues/ferreira/ricardo/auth/user/api/MyUserUpdatePasswordRequest.java
@@ -1,17 +1,0 @@
-package rodrigues.ferreira.ricardo.auth.user.api;
-
-import javax.validation.constraints.NotBlank;
-
-public class MyUserUpdatePasswordRequest {
-
-    @NotBlank
-    private String password;
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-}
