--- user/src/main/java/rodrigues/ferreira/ricardo/auth/user/api/ResourceNotFoundException.java
@@ -1,15 +1,0 @@
-package rodrigues.ferreira.ricardo.auth.user.api;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(HttpStatus.NOT_FOUND)
-public class ResourceNotFoundException extends RuntimeException {
-
-	public ResourceNotFoundException() {
-	}
-
-	public ResourceNotFoundException(String message) {
-		super(message);
-	}
-}
