+++ user/src/main/java/rodrigues/ferreira/ricardo/auth/user/api/exception/ResourceNotFoundException.java
@@ -1,0 +1,15 @@
+package rodrigues.ferreira.ricardo.auth.user.api.exception;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ResponseStatus(HttpStatus.NOT_FOUND)
+public class ResourceNotFoundException extends RuntimeException {
+
+	public ResourceNotFoundException() {
+	}
+
+	public ResourceNotFoundException(String message) {
+		super(message);
+	}
+}
