+++ user/src/main/java/rodrigues/ferreira/ricardo/auth/user/api/request/MyUserUpdatePasswordRequest.java
@@ -1,0 +1,14 @@
+package rodrigues.ferreira.ricardo.auth.user.api.request;
+
+import lombok.Getter;
+import lombok.Setter;
+
+import javax.validation.constraints.NotBlank;
+@Getter
+@Setter
+public class MyUserUpdatePasswordRequest {
+
+    @NotBlank
+    private String password;
+
+}
