+++ user/src/main/java/rodrigues/ferreira/ricardo/auth/user/domain/service/UserPasswordService.java
@@ -1,0 +1,77 @@
+package rodrigues.ferreira.ricardo.auth.user.domain.service;
+
+import lombok.RequiredArgsConstructor;
+import lombok.SneakyThrows;
+import org.springframework.security.core.token.KeyBasedPersistenceTokenService;
+import org.springframework.security.core.token.SecureRandomFactoryBean;
+import org.springframework.security.core.token.Token;
+import org.springframework.security.crypto.password.PasswordEncoder;
+import org.springframework.stereotype.Service;
+import rodrigues.ferreira.ricardo.auth.user.domain.PasswordTokenPublicData;
+import rodrigues.ferreira.ricardo.auth.user.domain.exception.UserEntityNotFoundException;
+import rodrigues.ferreira.ricardo.auth.user.domain.repository.UserRepository;
+import rodrigues.ferreira.ricardo.auth.user.domain.entity.UserEntity;
+
+import java.time.Duration;
+import java.time.Instant;
+import java.util.Base64;
+import java.util.Date;
+
+@Service
+@RequiredArgsConstructor
+public class UserPasswordService {
+
+    private final UserRepository userRepository;
+    private final PasswordEncoder passwordEncoder;
+
+    @SneakyThrows
+    public String generateToken(UserEntity user) {
+        KeyBasedPersistenceTokenService tokenService = getInstanceFor(user);
+
+        Token token = tokenService.allocateToken(user.getEmail());
+
+        return token.getKey();
+    }
+
+    @SneakyThrows
+    public void changePassword(String newPassword, String rawToken) {
+        PasswordTokenPublicData publicData = readPublicData(rawToken);
+
+        if(isExpired(publicData)) {
+            throw new RuntimeException("Expired token");
+        }
+
+        UserEntity userEntity = userRepository.findByEmail(publicData.getEmail())
+                .orElseThrow(UserEntityNotFoundException::new);
+
+        KeyBasedPersistenceTokenService tokenService = this.getInstanceFor(userEntity);
+        tokenService.verifyToken(rawToken);
+
+        userEntity.setPassword(this.passwordEncoder.encode(newPassword));
+        userRepository.save(userEntity);
+    }
+
+    private boolean isExpired(PasswordTokenPublicData publicData) {
+        Instant createdAt = new Date(publicData.getCreateAtTimestamp()).toInstant();
+        Instant now = new Date().toInstant();
+        return createdAt.plus(Duration.ofMinutes(5)).isBefore(now);
+    }
+
+    private KeyBasedPersistenceTokenService getInstanceFor(UserEntity user) throws Exception {
+        KeyBasedPersistenceTokenService tokenService = new KeyBasedPersistenceTokenService();
+
+        tokenService.setServerSecret(user.getPassword());
+        tokenService.setServerInteger(16);
+        tokenService.setSecureRandom(new SecureRandomFactoryBean().getObject());
+        return tokenService;
+    }
+
+    private PasswordTokenPublicData readPublicData(String rawToken) {
+        String rawTokenDecoded = new String(Base64.getDecoder().decode(rawToken));
+        String[] tokenParts = rawTokenDecoded.split(":");
+        Long timestamp = Long.parseLong(tokenParts[0]);
+        String email = tokenParts[2];
+        return new PasswordTokenPublicData(email, timestamp);
+    }
+
+}
