+++ user/src/main/java/rodrigues/ferreira/ricardo/auth/user/security/JPAUserDetailsService.java
@@ -1,6 +1,6 @@
 package rodrigues.ferreira.ricardo.auth.user.security;
 
-import rodrigues.ferreira.ricardo.auth.user.domain.UserRepository;
+import rodrigues.ferreira.ricardo.auth.user.domain.repository.UserRepository;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.User;
 import org.springframework.security.core.userdetails.UserDetails;
