+++ user/src/main/resources/application.yml
@@ -1,5 +1,5 @@
 server:
-  port: 8080
+  port: 8083
 spring:
   datasource:
     username: app_employees
